Modified emulate_hop in the Linux module so that, if tailexit is computed to be equal to curtick (which is possible in some cases if the link delay is 0) then we emulate_hop again.
This mitigates erroneous behavior that was previously occuring when
link delay was set to 0: a lot of packets where curtick == tailexit
weren't getting handled for several seconds, till the packet calendar
wrapped around again.

This change still needs to be verified for effects on packet reordering.
To be safe, it is preferable to avoid links with delay set to 0 milliseconds.